requirements.txt
einops
requests
pytest
click
joblib
tqdm
ipykernel
notebook
ipywidgets
jupyter
flask
matplotlib
numpy-stl
mediapy
pandas
scikit-learn
black
wandb==0.13.10
ftfy
tensorboard
plotly
sympy
accelerate
pre-commit
diffusers
gdown
jaxtyping
typeguard
streamlit
streamlit-analytics
torchinfo
frozendict
streamlit-chat
openai
transformer_lens
git+https://github.com/callummcdougall/CircuitsVis.git#subdirectory=python
git+https://github.com/neelnanda-io/neel-plotly
git+https://github.com/llm-attacks/llm-attacks
gym==0.23.1
pygame
autorom[accept-rom-license]
ale-py
protobuf==3.20.3